AMC/Engine: Revert changes to fadepal from commit 1a23b2de

+++ source/duke3d/src/screens.cpp
@@ -1401,7 +1401,7 @@ void G_DisplayRest(int32_t smoothratio)
     VM_OnEvent(EVENT_DISPLAYEND, g_player[screenpeek].ps->i, screenpeek);
 }
 
-void G_FadePalette(uint8_t r, uint8_t g, uint8_t b, uint8_t e)
+void G_FadePalette(int32_t r, int32_t g, int32_t b, int32_t e)
 {
     if (ud.screenfade == 0)
       return;
@@ -1415,7 +1415,7 @@ void G_FadePalette(uint8_t r, uint8_t g, uint8_t b, uint8_t e)
 
 // START and END limits are always inclusive!
 // STEP must evenly divide END-START, i.e. abs(end-start)%step == 0
-void fadepal(uint8_t r, uint8_t g, uint8_t b, uint8_t start, uint8_t end, uint8_t step)
+void fadepal(int32_t r, int32_t g, int32_t b, int32_t start, int32_t end, int32_t step)
 {
     if (ud.screenfade == 0)
       return;
@@ -1442,7 +1442,7 @@ void fadepal(uint8_t r, uint8_t g, uint8_t b, uint8_t start, uint8_t end, uint8_
 }
 
 // START and END limits are always inclusive!
-static void fadepaltile(uint8_t r, uint8_t g, uint8_t b, uint8_t start, uint8_t end, uint8_t step, uint16_t tile)
+static void fadepaltile(int32_t r, int32_t g, int32_t b, int32_t start, int32_t end, int32_t step, int32_t tile)
 {
     if (ud.screenfade == 0)
       return;
