+++ source/duke3d/src/screens.h
@@ -44,4 +44,4 @@ extern int32_t dr_yxaspect, dr_viewingrange;
 extern int32_t g_noLogoAnim, g_noLogo;
 extern double g_moveActorsTime, g_moveWorldTime;
 
-extern void G_FadePalette(uint8_t r, uint8_t g, uint8_t b, uint8_t e);
+extern void G_FadePalette(int32_t r, int32_t g, int32_t b, int32_t e);
